Added project dependencies: dotenv, dry-types.
Dry-Types will be helpful for defining value objects, form objects and
entities for the code.

Dotenv is a nice addition for 12-factor Rails apps - a convenient way of
storing development-mode environment variables. Using secrets.yml would
be also possible, but this way it is more conventional for devops which
are accustomed to set environment variables to provide necessary pieces
of information to an app.