Introduced TargetGroups::Hierarchy entity and repository object.
Target Groups have certain properties which makes it hard to express all
business properties in an ActiveRecord object - that's why it is viable
to provide an abstraction over it with the associated repository. This
way constraint like 'country can be linked only to a root target group'
is naturally enforced - there is no way in TargetGroups::Hierarchy
interface to create a wrong state.

Hierarchy object is also optimized to quick lookup of nodes - making the
interface fluent & easy to use and performant. The recreation of the
Hierarchy object is O(h) where h is a depth of the tree.